Safari-609.1.11.1/LayoutTests/platform/mac-mojave/http/tests/cookies/same-site/popup-same-site-via-cross-site-redirect-expected.txt


FAIL '127.0.0.1' is same-site with itself, so samesite cookies are sent. assert_equals: implicit-strict expected (string) "1" but got (undefined) undefined
